Get the acceleration structure and procedural geo working.
NOTE: currently we see the silhouette of a cube, yet the geometry is a plane.
Our intersection shader is doing nothing wrong, it is just that our AABB is too
small. the rays are intersecting with a "further back" part of the plane. and hence
the silhouette is really just an artifact of a cube silhouette "cookie cutter".