Implement GLTF loading.
There are still some artefacts in the produced image.
I think this is because I lowered the bounce count
and the rays per pixel. I suspect we would win if
we tune em' back up again.

Next to step is optimization. I benchmarked the render
time for our image: it's 165.93 seconds. I think we can
make this faster :)